requirements.txt
asgiref==3.4.1
boto3==1.20.9
botocore==1.23.9
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
mysqlclient==2.1.0
Pillow==8.4.0
psycopg2==2.7.7
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
whitenoise==5.3.0
